🎨 동일 UI 구조 변경 및 레이아웃 제작
